log conf moved to config dir
