+++ tests/test_logging_conf.py
@@ -2,7 +2,7 @@
 Tests logging_config.py module
 """
 
-from nightshift.logging_conf import get_handlers, get_token, log_config
+from nightshift.config.logging_conf import get_handlers, get_token, log_conf
 
 
 def test_get_handlers(mock_log_env):
@@ -13,8 +13,8 @@ def test_get_token(mock_log_env):
     assert get_token() == "ns_token_here"
 
 
-def test_log_config(mock_log_env):
-    conf = log_config()
+def test_log_conf(mock_log_env):
+    conf = log_conf()
     assert sorted(conf.keys()) == [
         "disable_existing_loggers",
         "formatters",
